Adding the Object File Library
This Library must be linked in your project.

In the Solution Explorer, right click on your console application name.

In the quick menu, select properties 

In the property pages , go to linker and then input.

In additional dependencies add the path to the .lib object.  Like this, "<pathTo>\RPPLibrary.lib"

Click Apply to save changes then Ok.